chore: changes for rebase on 0.29
